backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
# Updated LangChain packages with consistent versions
langchain>=0.0.267
langchain-experimental>=0.0.35
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-openai>=0.0.2
# Updated OpenAI version to support latest API
openai>=1.0.0
neo4j>=5.9.0
python-dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.12.2
PyPDF2>=3.0.1
docx2txt>=0.8
pytest>=7.3.1
json-repair>=0.5.0
# AI model integrations
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.1